Move custom metric classes into a subpackage.
