+++ spark/v3.2/spark/src/main/java/org/apache/iceberg/spark/source/metrics/NumSplits.java
@@ -17,16 +17,12 @@
  * under the License.
  */
 
-package org.apache.iceberg.spark.source;
+package org.apache.iceberg.spark.source.metrics;
 
-import org.apache.spark.sql.connector.metric.CustomTaskMetric;
+import java.text.NumberFormat;
+import org.apache.spark.sql.connector.metric.CustomMetric;
 
-public class TaskNumSplits implements CustomTaskMetric {
-  private long value;
-
-  TaskNumSplits(long value) {
-    this.value = value;
-  }
+public class NumSplits implements CustomMetric {
 
   @Override
   public String name() {
@@ -34,7 +30,16 @@ public class TaskNumSplits implements CustomTaskMetric {
   }
 
   @Override
-  public long value() {
-    return value;
+  public String description() {
+    return "number of file splits read";
+  }
+
+  @Override
+  public String aggregateTaskMetrics(long[] taskMetrics) {
+    long sum = initialValue;
+    for (int i = 0; i < taskMetrics.length; i++) {
+      sum += taskMetrics[i];
+    }
+    return NumberFormat.getIntegerInstance().format(sum);
   }
 }
